Handle closed/invalid session handles.
There are three cases in which a session can become unusable.
See:
https://www.cryptsoft.com/pkcs11doc/v200/group__SEC__9__FUNCTIONS.html,
specifically the section "Cryptoki function return values for functions
that use a session handle".
1. The session handle was made invalid (because a token was removed or
the session was closed by this or another application with knowledge of
the handle).
2. A device was removed from its slot.
3. The session was closed during the execution of the function.
In any of these cases, a sane response is to reopen the slot, and
attempt to make the call again.